refactor(version): single source of truth for akm-cli version contract + accept 0.9.x
The required-version range lived in three places with three syntaxes — the
Claude hook constant, the OpenCode constant string, and a hand-rolled
`minor === 8` integer matcher in opencode that had drifted from both. Bumping
the range meant editing all three.

- Add claude/shared/akm-version.ts: exports AKM_VERSION_RANGE and
  satisfiesAkmVersionRange(), the latter routing through the vendored semver
  matcher (the Claude hook runs as a bare Bun script with no node_modules, so
  it can't use the npm `semver` package).
- Move vendor-semver.ts into claude/shared/ so both the hook and the vendored
  OpenCode tarball share one matcher.
- akm-hook.ts and opencode/index.ts now import the shared range; deleted the
  duplicate constant and the hardcoded minor===8 check.
- Widen the range to accept 0.9.x including prereleases (^0.9.0-beta.0 || ^0.9.0)
  so 0.9.0-beta.6 satisfies the gate.

OpenCode is now consistent with (and as strict as) the Claude hook: patch-level
prereleases like 0.8.1-rc0 that the old integer matcher accepted are now
rejected by real semver, matching the hook. All published versions
(0.8.0-rc.N, 0.8.x, 0.9.0-beta.N, 0.9.x) match on both sides.

prepack vendoring carries both new shared files into the standalone OpenCode
tarball; postpack reverts cleanly. 326 tests pass.
