+++ claude/shared/vendor-semver.ts
